pyproject.toml
[project]
name = "learning-tree"
version = "0.1.0"
description = "A simple Flask web application for tracking your learning journey in a tree-like structure"
requires-python = ">=3.8"
dependencies = [
    "flask==3.0.2",
    "flask-sqlalchemy==3.1.1",
    "python-dateutil==2.8.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 